tests: send messages at a certain frequency
The two publishers must be launched at the same time, for example with:

$ ./test_publisher ... & ./test_publisher ...
^C
$ fg
^C

The subscriber prints the first and last messages received for each
publisher. With the message numbers, we can see that the messages are
kept in the queue (or not). With 100Hz during 3 seconds, all messages
are kept in the queue for the second publisher, and when we ask to
receive the messages (of the second publisher), we get all of them
directly without waiting. So, it works. But there is a memory limit. It
is most probably possible to configure the memory limit (high-water
mark, or something), but some tests should be done to check that we will
receive all messages from the Pupil (60 Hz during 30 seconds, minimum,
better check for 100 Hz during 2 minutes, to be sure).