Update docs
The doc is not written at the same place as the code, so we often forget
the documentation, it's a bad thing. When the documenation is close to
the related code, there are more chances to update the doc directly, and
write it directly too. Matlab is not great at that.